+++ .idea/encodings.xml
@@ -1,6 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/tokens/StoredCredential.json" charset="ISO-8859-1" />
-  </component>
+  <component name="Encoding" defaultCharsetForPropertiesFiles="UTF-8" />
 </project>
