+++ src/main/java/DataStructures/SimplifiedEvent.java
@@ -1,0 +1,17 @@
+package DataStructures;
+
+public class SimplifiedEvent {
+    public String title;
+    public String start;
+    public String end;
+    public String location;
+    public String description;
+
+    public SimplifiedEvent(String title, String start, String end, String location, String description) {
+        this.title = title;
+        this.start = start;
+        this.end = end;
+        this.location = location;
+        this.description = description;
+    }
+}
