+++ src/main/resources/language/English/EventPrompt.json
@@ -4,5 +4,5 @@
   "Review": "The following json format data is my [target event], please help me to arrange the learning schedule.",
   "Curriculum": "The following csv format data is my [curriculum].",
   "Project": "The following json format data is my [target event],please help me to arrange the working schedule.",
-  "End": "and return only the json format data containing [title], [start], [end], [location], [description] in list.To notice that the [original events] should [not] be included and time format should be 3999."
+  "End": "and return only the json format data containing [title], [start], [end], [location], [description] in list.To notice that the [original events] should [not] be included and time format should yyyy-MM-dd'T'HH:mm:ssXXX."
 }
